non-working comimit. Mistake: Made edits to BuildCostVolume(..) instead of BuildCostVolume2(..), for integration of cam2cam transform matrix with inverse depth cost vol.

+++ codes/DtamRealizition/fileLoader.hpp
@@ -4,7 +4,7 @@
 // loads all files of a given name and extension
 #include "convertAhandaPovRayToStandard.h"
 void loadAhanda(const char * rootpath,
-                double range,
+                float range,
                 int imageNumber,
                 cv::Mat& image,
                 cv::Mat& d,
